Remove old footer; rename LandingPage; add ComponentTest for footer
