+++ dataland-frontend/src/components/pages/AboutPage.vue
@@ -4,14 +4,14 @@
     <TheAboutIntro :sections="aboutPage?.sections" />
     <TheAboutPrinciples :sections="aboutPage?.sections" />
   </main>
-  <TheFooter :sections="landingPage?.sections" :is-light-version="true" />
+  <TheFooter />
 </template>
 
 <script setup lang="ts">
 import TheHeader from '@/components/generics/TheNewHeader.vue';
 import TheAboutIntro from '@/components/resources/aboutPage/TheAboutIntro.vue';
 import TheAboutPrinciples from '@/components/resources/aboutPage/TheAboutPrinciples.vue';
-import TheFooter from '@/components/generics/TheNewFooter.vue';
+import TheFooter from '@/components/generics/TheFooter.vue';
 import contentData from '@/assets/content.json';
 import type { Content, Page } from '@/types/ContentTypes';
