+++ dataland-frontend/src/components/pages/DatasetOverview.vue
@@ -20,26 +20,24 @@
         </div>
       </DatasetsTabMenu>
     </TheContent>
-    <TheFooter :is-light-version="true" :sections="footerContent" />
+    <TheFooter />
   </AuthenticationWrapper>
 </template>
 
 <script lang="ts">
-import AuthenticationWrapper from '@/components/wrapper/AuthenticationWrapper.vue';
-import TheHeader from '@/components/generics/TheHeader.vue';
-import TheContent from '@/components/generics/TheContent.vue';
-import { defineComponent, inject } from 'vue';
-import TheFooter from '@/components/generics/TheNewFooter.vue';
-import contentData from '@/assets/content.json';
-import type { Content, Page } from '@/types/ContentTypes';
-import DatasetOverviewTable from '@/components/resources/datasetOverview/DatasetOverviewTable.vue';
-import { assertDefined } from '@/utils/TypeScriptUtils';
-import type Keycloak from 'keycloak-js';
-import { type DatasetTableInfo, getMyDatasetTableInfos } from '@/components/resources/datasetOverview/DatasetTableInfo';
-import { checkIfUserHasRole } from '@/utils/KeycloakUtils';
 import DatasetsTabMenu from '@/components/general/DatasetsTabMenu.vue';
 import NewDatasetButton from '@/components/general/NewDatasetButton.vue';
+import TheContent from '@/components/generics/TheContent.vue';
+import TheFooter from '@/components/generics/TheFooter.vue';
+import TheHeader from '@/components/generics/TheHeader.vue';
+import DatasetOverviewTable from '@/components/resources/datasetOverview/DatasetOverviewTable.vue';
+import { type DatasetTableInfo, getMyDatasetTableInfos } from '@/components/resources/datasetOverview/DatasetTableInfo';
+import AuthenticationWrapper from '@/components/wrapper/AuthenticationWrapper.vue';
 import { KEYCLOAK_ROLE_UPLOADER } from '@/utils/KeycloakRoles';
+import { checkIfUserHasRole } from '@/utils/KeycloakUtils';
+import { assertDefined } from '@/utils/TypeScriptUtils';
+import type Keycloak from 'keycloak-js';
+import { defineComponent, inject } from 'vue';
 
 export default defineComponent({
   name: 'DatasetOverview',
@@ -53,14 +51,10 @@ export default defineComponent({
     DatasetOverviewTable,
   },
   data() {
-    const content: Content = contentData;
-    const footerPage: Page | undefined = content.pages.find((page) => page.url === '/');
-    const footerContent = footerPage?.sections;
     return {
       datasetTableInfos: [] as DatasetTableInfo[],
       waitingForData: true,
       hasUserUploaderRights: undefined as boolean | undefined,
-      footerContent,
     };
   },
   setup() {
