+++ dataland-frontend/src/components/pages/LandingPage.vue
@@ -9,7 +9,7 @@
     <TheJoinCampaign :sections="landingPage?.sections" />
     <TheGetInTouch :sections="landingPage?.sections" />
   </main>
-  <TheFooter :sections="landingPage?.sections" :isLightVersion="false" />
+  <TheFooter :isLightVersion="false" />
 </template>
 
 <script setup lang="ts">
@@ -27,7 +27,7 @@ import TheStruggle from '@/components/resources/newLandingPage/TheStruggle.vue';
 import TheHowItWorks from '@/components/resources/newLandingPage/TheHowItWorks.vue';
 import TheJoinCampaign from '@/components/resources/newLandingPage/TheJoinCampaign.vue';
 import TheGetInTouch from '@/components/resources/newLandingPage/TheGetInTouch.vue';
-import TheFooter from '@/components/generics/TheNewFooter.vue';
+import TheFooter from '@/components/generics/TheFooter.vue';
 import contentData from '@/assets/content.json';
 import type { Content, Page } from '@/types/ContentTypes';
