+++ dataland-frontend/src/components/pages/UnsubscribeFromMailsPage.vue
@@ -23,26 +23,24 @@
     </main>
   </TheContent>
 
-  <TheNewFooter :is-light-version="true" :sections="footerPageSections" class="footer" />
+  <TheFooter />
 </template>
 
 <script lang="ts">
-import { defineComponent, inject } from 'vue';
-import PrimeButton from 'primevue/button';
-import type Keycloak from 'keycloak-js';
-import TheHeader from '@/components/generics/TheHeader.vue';
 import TheContent from '@/components/generics/TheContent.vue';
-import TheNewFooter from '@/components/generics/TheNewFooter.vue';
+import TheFooter from '@/components/generics/TheFooter.vue';
+import TheHeader from '@/components/generics/TheHeader.vue';
 import { ApiClientProvider } from '@/services/ApiClients';
 import { assertDefined } from '@/utils/TypeScriptUtils';
-import type { Content, Page } from '@/types/ContentTypes';
-import contentData from '@/assets/content.json';
+import type Keycloak from 'keycloak-js';
+import PrimeButton from 'primevue/button';
+import { defineComponent, inject } from 'vue';
 
 export default defineComponent({
   name: 'UnsubscribeFromMailsPage',
   components: {
+    TheFooter,
     TheContent,
-    TheNewFooter,
     TheHeader,
     PrimeButton,
   },
@@ -61,10 +59,7 @@ export default defineComponent({
   },
 
   data() {
-    const content: Content = contentData;
-    const footerPage: Page | undefined = content.pages.find((page) => page.url === '/');
     return {
-      footerPageSections: footerPage?.sections,
       isUnsubscribed: false,
       message: '',
     };
