+++ dataland-frontend/tests/component/components/pages/LandingPage.cy.ts
@@ -1,14 +1,15 @@
 import { checkButton, checkImage, checkAnchorByContent } from '@ct/testUtils/ExistenceChecks';
-import NewLandingPage from '@/components/pages/NewLandingPage.vue';
+import LandingPage from '@/components/pages/LandingPage.vue';
 import { minimalKeycloakMock } from '@ct/testUtils/Keycloak';
 import content from '@/assets/content.json';
 import { type Page, type Section } from '@/types/ContentTypes';
 import { assertDefined } from '@/utils/TypeScriptUtils';
+import { checkFooter } from '@sharedUtils/ElementChecks.ts';
 import { setMobileDeviceViewport } from '@sharedUtils/TestSetupUtils';
 
 describe('Component test for the landing page', () => {
   it('Check if essential elements are present', () => {
-    cy.mountWithPlugins(NewLandingPage, {
+    cy.mountWithPlugins(LandingPage, {
       keycloak: minimalKeycloakMock({
         authenticated: false,
       }),
@@ -26,7 +27,7 @@ describe('Component test for the landing page', () => {
       assertFrameworkPanelExists('SFDR');
       cy.get('button.joincampaign__button').should('exist');
       cy.get('button.getintouch__text-button').should('exist');
-      checkNewFooter();
+      checkFooter();
 
       setMobileDeviceViewport();
       validateTheHeader();
@@ -65,30 +66,6 @@ function validateBrandsSection(): void {
   images?.forEach((image, index) => {
     const filename = image.split('/').slice(-1)[0];
     checkImage(`Brand ${index + 1}`, filename);
-  });
-}
-
-/**
- * Check the new footer for long-term sustainability.
- */
-function checkNewFooter(): void {
-  cy.get('footer').should('exist');
-
-  cy.get('.footer__logo').should('exist');
-
-  const currentYear = new Date().getFullYear();
-  const expectedCopyrightText = `Copyright © ${currentYear} Dataland`;
-
-  cy.get('.footer__copyright').should('contain.text', expectedCopyrightText);
-
-  const essentialLinks = [
-    { href: '/imprint', text: 'IMPRINT' },
-    { href: '/dataprivacy', text: 'DATA PRIVACY' },
-    { href: '/terms', text: 'LEGAL' },
-  ];
-
-  essentialLinks.forEach((link) => {
-    cy.get(`footer a[href='${link.href}']`).should('contain.text', link.text);
   });
 }
